Rename: 컴포넌트 이름 적절하게 변경
